Added BaseRobot
