+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/EncoderAuto.java
@@ -53,7 +53,7 @@ import com.qualcomm.robotcore.util.Range;
 
 @TeleOp(name="Basic: Linear OpMode", group="Linear Opmode")
 @Disabled
-public class EncoderAuto extends LinearOpMode {
+public class EncoderAuto extends BaseRobot {
 
     private double TICKS_PER_ROTATION=1500;
     private double INCHES_PER_ROTATION=3; //wheel's circumference
@@ -120,47 +120,7 @@ public class EncoderAuto extends LinearOpMode {
         }
     }
 
-    public void mecanum (double y, double x, double rx) {
-        double frontLeftPower=y + x + rx;
-        double backLeftPower=y - x + rx;
-        double frontRightPower=y - x - rx;
-        double backRightPower=y + x - rx;
-        if (Math.abs(frontLeftPower) > 1 || Math.abs(backLeftPower) > 1 ||
-                Math.abs(frontRightPower) > 1 || Math.abs(backRightPower) > 1 ) {
-            // Find the largest power
-            double max = 0;
-            max = Math.max(Math.abs(frontLeftPower), Math.abs(backLeftPower));
-            max = Math.max(Math.abs(frontRightPower), max);
-            max = Math.max(Math.abs(backRightPower), max);
 
-            // Divide everything by max (it's positive so we don't need to worry
-            // about signs)
-            frontLeftPower /= max;
-            backLeftPower /= max;
-            frontRightPower /= max;
-            backRightPower /= max;
-        }
-        frontLeftDrive.setPower(frontLeftPower);
-        backLeftDrive.setPower(backLeftPower);
-        frontRightDrive.setPower(frontRightPower);
-        backRightDrive.setPower(backRightPower);
-    }
-
-    public void reset_encoders() {
-        frontLeftDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        backLeftDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        backRightDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        frontRightDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-        all_run_using_encoders();
-    }
-
-    public void all_run_using_encoders() {
-        backLeftDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        backRightDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        frontRightDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        frontLeftDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-    }
